+++ spec/helpers/env.js
@@ -1,1 +1,3 @@
+const { configuration } = require("../../src");
+configuration.verbose = true;
 process.env.VUE_DI_TESTING = true;