+++ src/classification/ast.parser.ts
@@ -1,6 +1,8 @@
-import { SourceFile, VariableDeclaration, Node, ObjectLiteralExpression, SyntaxKind, SyntaxList, Identifier } from "ts-simple-ast";
+import { SourceFile, VariableDeclaration, Node, ObjectLiteralExpression, 
+    SyntaxKind, SyntaxList, Identifier, MemberExpression, StringLiteral } from "ts-simple-ast";
 import { Argumenter } from "@joejukan/argumenter";
 import { log } from "../function";
+import { ok } from "@joejukan/web-kit";
 
 export class ASTParser {
     private source: SourceFile;
@@ -47,14 +49,27 @@ export class ASTParser {
         }
         return undefined;
     }
-    public exists(key: string, node: Node): boolean{
-        return typeof this.get(key, node) !== 'undefined';
-    }
 
-    public get(key: string, node: Node): ObjectLiteralExpression
-    public get(...args): ObjectLiteralExpression{
+    public exists(key: string, node: Node)
+    public exists(...args): boolean{
         let argue = new Argumenter(args);
         let key = argue.string;
+        let kind: SyntaxKind = argue.number;
+        let node:Node = <Node> argue.object;
+
+        return typeof this.get(node, key, <any> kind) !== 'undefined';
+    }
+
+    public get(node: Node, key: string): MemberExpression;
+    public get(key: string, node: Node): MemberExpression;
+    public get(node: Node, key: string, kind: SyntaxKind.StringLiteral): StringLiteral;
+    public get(key: string, node: Node, kind: SyntaxKind.StringLiteral): StringLiteral;
+    public get(node: Node, key: string, kind: SyntaxKind.ObjectLiteralExpression): ObjectLiteralExpression;
+    public get(key: string, node: Node, kind: SyntaxKind.ObjectLiteralExpression): ObjectLiteralExpression;
+    public get(...args): MemberExpression {
+        let argue = new Argumenter(args);
+        let key = argue.string;
+        let kind: SyntaxKind = argue.number;
         let node:Node = <Node> argue.object;
         let syntax:SyntaxList = undefined;
 
@@ -68,12 +83,17 @@ export class ASTParser {
         }
 
         if(syntax){
-            let pairs = syntax.getChildren()
-            pairs.forEach(pair => {
+            let pairs = syntax.getChildren() || [];
+            for(let i = 0; i < pairs.length; i++){
+                let pair = pairs[0];
                 let id = pair.getFirstChildByKind(SyntaxKind.Identifier);
-                if(key === id.getText() || `'${key}'` === id.getText() || `"${key}"` === id.getText())
-                    return pair.getFirstChildByKind(SyntaxKind.ObjectLiteralExpression);
-            });
+                let literal = <MemberExpression> ( ok(kind) ? pair.getFirstChildByKind(kind) : pair.getLastChild());
+                
+                if(key === id.getText() || `'${key}'` === id.getText() || `"${key}"` === id.getText()){
+                    return literal;
+                }
+                    
+            }
         }
 
         return